feat: implement built-in VPN, LLM search, and Aletheia OS integration
Built-in VPN (WireGuard-style):
- X25519 ECDH key exchange via OpenSSL 3.x
- ChaCha20-Poly1305 AEAD packet encryption
- HMAC-SHA256/HKDF-SHA256 for indexing and key derivation
- Full handshake state machine with forward secrecy
- CIDR-based routing, MTU enforcement, tamper detection
- Integrated into HttpClient for automatic traffic routing

LLM Search Service:
- Web search API for the Aletheia OS LLM agent
- Page reading with readable text extraction
- VPN-routed by default for privacy
- Structured results (title, URL, snippet, relevance)

Aletheia OS Bridge:
- Native browser control (open, navigate, read)
- VPN control from the OS
- LLM search integration
- Status reporting

Also:
- Fixed broken http_client.cc (syntax errors, literal newlines)
- Fixed sandbox.cc undefined symbol
- Rewrote test suite with lightweight framework (37 tests, all passing)
- Fixed CMakeLists for real paths and proper library structure

+++ include/config.h
@@ -5,6 +5,7 @@
 #include <string>
 #include <vector>
 #include <map>
+#include <cstdint>
 
 namespace lethe {
 
@@ -21,18 +22,18 @@ const bool ENFORCE_STRICT_CSP = true;
 
 const int MAX_OPEN_CONNECTIONS = 128;
 const int MAX_CONNECTIONS_PER_HOST = 16;
-const size_t NETWORK_BUFFER_SIZE = 65536;
+const uint64_t NETWORK_BUFFER_SIZE = 65536;
 
-const size_t MAX_MEMORY_PER_PROCESS = 1ULL << 30;
-const size_t TOTAL_MAX_MEMORY = 4ULL << 30;
+const uint64_t MAX_MEMORY_PER_PROCESS = 1ULL << 30; // 1GB
+const uint64_t TOTAL_MAX_MEMORY = 4ULL << 30; // 4GB total
 
 const std::vector<std::string> DOH_PROVIDERS = {
     "https://cloudflare-dns.com/dns-query",
     "https://dns.quad9.net/dns-query"
 };
 
-const int MIN_TLS_VERSION = 0x0304;
-const int MAX_TLS_VERSION = 0x0305;
+const int MIN_TLS_VERSION = 0x0304; // TLS 1.2+
+const int MAX_TLS_VERSION = 0x0305; // TLS 1.3
 const std::string DEFAULT_CIPHERS = 
     "TLS_AES_256_GCM_SHA384:TLS_CHACHA20_POLY1305_SHA256:"
     "TLS_AES_128_GCM_SHA256";
@@ -54,4 +55,4 @@ const std::string DEFAULT_CSP_POLICY =
 
 } // namespace lethe
 
-#endif
+#endif // LETHE_CONFIG_H
